[refactor/#6] - 아이콘 파일명 케밥케이스 변환 (ic_* → name-icon)
